+++ src/index.ts
@@ -1,3 +1,3 @@
-import DatasetJson from "./class/datasetJson";
+import DatasetJson from './class/datasetJson';
 export default DatasetJson;
 export * from "./interfaces/datasetJson";