+++ test/getData.test.ts
@@ -1,8 +1,8 @@
-import DatasetJson from "../src/index";
+import DatasetJson from '../src/index';
 
 
-test("Get 5 rows of dataset", async () => {
-    const filePath = "test/data/adsl.json";
+test('Get 5 rows of dataset', async () => {
+    const filePath = 'test/data/adsl.json';
 
     const data = new DatasetJson(filePath);
     const rows = await data.getData({start: 20, length: 5});
@@ -10,34 +10,34 @@ test("Get 5 rows of dataset", async () => {
     expect(rows).toMatchSnapshot();
 });
 
-test("Get 5 rows of dataset as an object", async () => {
-    const filePath = "test/data/adsl.json";
+test('Get 5 rows of dataset as an object', async () => {
+    const filePath = 'test/data/adsl.json';
 
     const data = new DatasetJson(filePath);
-    const rows = await data.getData({start: 20, length: 5, type: "object"});
+    const rows = await data.getData({start: 20, length: 5, type: 'object'});
     expect(rows.length).toBe(5);
     expect(rows).toMatchSnapshot();
 });
 
-test("Get 5 rows of dataset as an object and keep only specific variables", async () => {
-    const filePath = "test/data/adsl.json";
+test('Get 5 rows of dataset as an object and keep only specific variables', async () => {
+    const filePath = 'test/data/adsl.json';
 
     const data = new DatasetJson(filePath);
-    const rows = await data.getData({start: 20, length: 5, type: "object", filterColumns: ["USUBJID", "TRTEDT", "STUDYID"]});
+    const rows = await data.getData({start: 20, length: 5, type: 'object', filterColumns: ['USUBJID', 'TRTEDT', 'STUDYID']});
     expect(rows.length).toBe(5);
     expect(rows).toMatchSnapshot();
 });
 
-test("Get one row of NDJSON dataset", async () => {
-    const filePath = "test/data/adsl.ndjson";
+test('Get one row of NDJSON dataset', async () => {
+    const filePath = 'test/data/adsl.ndjson';
 
     const data = new DatasetJson(filePath);
     const rows = await data.getData({start: 0, length: 1});
     expect(rows.length).toBe(1);
     expect(rows).toMatchSnapshot();
 });
-test("Get 5 rows of NDJSON dataset", async () => {
-    const filePath = "test/data/adsl.ndjson";
+test('Get 5 rows of NDJSON dataset', async () => {
+    const filePath = 'test/data/adsl.ndjson';
 
     const data = new DatasetJson(filePath);
     const rows = await data.getData({start: 20, length: 5});
@@ -45,20 +45,20 @@ test("Get 5 rows of NDJSON dataset", async () => {
     expect(rows).toMatchSnapshot();
 });
 
-test("Get 5 rows of NDJSON dataset as an object", async () => {
-    const filePath = "test/data/adsl.ndjson";
+test('Get 5 rows of NDJSON dataset as an object', async () => {
+    const filePath = 'test/data/adsl.ndjson';
 
     const data = new DatasetJson(filePath);
-    const rows = await data.getData({start: 20, length: 5, type: "object"});
+    const rows = await data.getData({start: 20, length: 5, type: 'object'});
     expect(rows.length).toBe(5);
     expect(rows).toMatchSnapshot();
 });
 
-test("Get 5 rows of NDJSON dataset as an object and keep only specific variables", async () => {
-    const filePath = "test/data/adsl.ndjson";
+test('Get 5 rows of NDJSON dataset as an object and keep only specific variables', async () => {
+    const filePath = 'test/data/adsl.ndjson';
 
     const data = new DatasetJson(filePath);
-    const rows = await data.getData({start: 20, length: 5, type: "object", filterColumns: ["USUBJID", "TRTEDT", "STUDYID"]});
+    const rows = await data.getData({start: 20, length: 5, type: 'object', filterColumns: ['USUBJID', 'TRTEDT', 'STUDYID']});
     expect(rows.length).toBe(5);
     expect(rows).toMatchSnapshot();
 });